+++ db/migrate/20160824143346_add_qty_to_inventories.rb
@@ -1,5 +1,5 @@
 class AddQtyToInventories < ActiveRecord::Migration
   def change
-    add_column :inventories, :qty, :float
+    add_column :inventories, :qty, :integer
   end
 end
